backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose
bcrypt==4.1.2
python-multipart
